+++ delayed_image/__init__.py
@@ -284,7 +284,7 @@ Example:
 """
 
 
-__version__ = '0.3.1'
+__version__ = '0.4.0'
 __author__ = 'Jon Crall'
 __author_email__ = 'jon.crall@kitware.com'