+++ tests/test_off_by_one.py
@@ -1,16 +1,14 @@
 
 def test_off_by_one_with_upscale():
     import delayed_image
-    import numpy as np
     delayed = delayed_image.DelayedLoad.demo(key='astro').prepare()
-
-    orig = delayed.finalize()
 
     x = delayed.scale(2.0001)[0:4, 0:4]
     data1 = x.finalize(optimize=False, nodata_method='float')
     data2 = x.finalize(nodata_method='float')
-    assert np.all(data2 == data1)
 
+    import kwarray
+    assert kwarray.ArrayAPI.coerce('numpy').array_equal(data2, data1, equal_nan=True)
     if 0:
         import kwplot
         kwplot.autompl()
@@ -23,13 +21,14 @@ def test_off_by_one_with_multi_scale():
     import delayed_image
     import numpy as np
     delayed = delayed_image.DelayedLoad.demo(key='astro').prepare()
-    orig = delayed.finalize()
 
     for s in np.linspace(0.4, 2.5, 100).tolist() + [0.5, 1.0, 2.0]:
         x = delayed.scale(s)[0:4, 0:4]
         data1 = x.finalize(optimize=False, nodata_method='float')
         data2 = x.finalize(nodata_method='float')
-        assert np.all(data2 == data1)
+        # assert np.all(data2 == data1)
+        import kwarray
+        assert kwarray.ArrayAPI.coerce('numpy').array_equal(data2, data1, equal_nan=True)
 
         # import kwplot
         # kwplot.autompl()
@@ -40,13 +39,16 @@ def test_off_by_one_with_multi_scale():
 
 def test_off_by_one_with_small_img():
     """
-    This doesn't work right because of the align corners issue
+    This originally did not work correctly because of warp was using
+    integer-centers instead of integer-corners. The issue is described well in
+    this article [WhereArePixels]_.
 
     References:
         .. [ResizeConfusion] https://jricheimer.github.io/tensorflow/2019/02/11/resize-confusion/
         .. [InvWarp] https://github.com/ClementPinard/SfmLearner-Pytorch/blob/master/inverse_warp.py
         .. [TorchAffineTransform] https://discuss.pytorch.org/t/affine-transformation-matrix-paramters-conversion/19522
         .. [TorchIssue15386] https://github.com/pytorch/pytorch/issues/15386
+        .. [WhereArePixels] https://ppwwyyxx.com/blog/2021/Where-are-Pixels/
         # https://github.com/pytorch/pytorch/issues/20785
         # https://github.com/pytorch/pytorch/pull/23923
         # https://github.com/pytorch/pytorch/pull/24929
@@ -57,9 +59,6 @@ def test_off_by_one_with_small_img():
         ~/code/kwimage/kwimage/im_cv2.py
 
     """
-    import pytest
-    pytest.skip('This is broken')
-
     import delayed_image
     import kwimage
     import numpy as np
@@ -72,7 +71,7 @@ def test_off_by_one_with_small_img():
 
     warp = kwimage.Affine.coerce(offset=(0, 0), scale=(8.6, 8.5))
     x = delayed.warp(warp)
-    data = x.finalize(interpolation='nearest')
+    data1 = x.finalize(interpolation='nearest')
 
     warp = kwimage.Affine.coerce(offset=(0, 0), scale=(2, 2))
     warp = warp @ kwimage.Affine.translate((0.5, 0.5))
@@ -81,12 +80,31 @@ def test_off_by_one_with_small_img():
 
     data3 = kwimage.imresize(raw, scale=2.0, interpolation='nearest')
 
-    if 1:
+    SHOW = 0
+    if SHOW:
         import kwplot
         pnum_ = kwplot.PlotNums(nRows=1, nCols=4)
         kwplot.autompl()
         kwplot.plt.ion()
-        kwplot.imshow(raw, pnum=pnum_())
-        kwplot.imshow(kwimage.fill_nans_with_checkers(data), pnum=pnum_())
-        kwplot.imshow(kwimage.fill_nans_with_checkers(data2), pnum=pnum_())
-        kwplot.imshow(kwimage.fill_nans_with_checkers(data3), pnum=pnum_())
+        kwplot.imshow(raw, pnum=pnum_(), title='original image', show_ticks=True, origin_convention='corner')
+        kwplot.imshow(kwimage.fill_nans_with_checkers(data1.copy()), pnum=pnum_(), title='scaled by non-integer', show_ticks=True, origin_convention='corner')
+        kwplot.imshow(kwimage.fill_nans_with_checkers(data2.copy()), pnum=pnum_(), title='scaled by 2 and shifted by 0.5', show_ticks=True, origin_convention='corner')
+        kwplot.imshow(kwimage.fill_nans_with_checkers(data3.copy()), pnum=pnum_(), title='imresize scale by 2', show_ticks=True, origin_convention='corner')
+
+    raw.shape
+    assert np.all(np.unique(raw) == np.unique(data1)), (
+        'data1 should have exactly the same values as raw because it is '
+        'just an upscale with nearest resampling. '
+        'It should not have any nan values')
+
+    assert not np.any(np.isnan(data2[1:, 1:])), (
+        'data2 should not have any nan values except in the first row / column '
+        'due to the 0.5 translation')
+
+    assert np.all(np.isnan(data2[:1, :])), (
+        'data2 first row should be all nans due to a shift by 0.5 and scale by 2')
+    assert np.all(np.isnan(data2[:, :1])), (
+        'data2 first row should be all nans due to a shift by 0.5 and scale by 2')
+
+    assert not np.any(np.isnan(data3)), (
+        'data3 is just a sanity check and should not have nans due to imresize implementation')